feat(tui): rewrite runtime around crossterm modules
