+++ Cargo.toml
@@ -23,7 +23,7 @@ sha2 = "0.10"
 urlencoding = "2.1"
 url = "2.5"
 ratatui = "0.30"
-crossterm = "0.29"
+crossterm = { version = "0.29", features = ["event-stream"] }
 walkdir = "2.5"
 glob = "0.3"
 regex = "1.11"
